Expose AppDomain.PermissionSet
