+++ src/System.Runtime.Extensions/src/System.Runtime.Extensions.csproj
@@ -53,6 +53,9 @@
     <Compile Include="System\Security\Permissions\SecurityAttribute.cs" />
     <Compile Include="System\Security\Permissions\SecurityPermissionAttribute.cs" />
     <Compile Include="System\Security\Permissions\SecurityPermissionFlag.cs" />
+    <Compile Include="..\..\System.Security.Permissions\src\System\Security\IStackWalk.cs" />
+    <Compile Include="..\..\System.Security.Permissions\src\System\Security\PermissionSet.cs" />
+    <Compile Include="..\..\System.Security.Permissions\src\System\Security\Permissions\PermissionState.cs" />
     <Compile Include="$(CommonPath)\CoreLib\System\IO\StreamHelpers.CopyValidation.cs">
       <Link>Common\CoreLib\System\IO\StreamHelpers.CopyValidation.cs</Link>
     </Compile>
