+++ src/System.Security.AccessControl/ref/Configurations.props
@@ -3,6 +3,7 @@
   <PropertyGroup>
     <BuildConfigurations>
       netstandard;
+      netcoreapp;
       netfx;
       net461;
     </BuildConfigurations>
