+++ src/System.Security.AccessControl/ref/System.Security.AccessControl.csproj
@@ -17,4 +17,8 @@
     <SuppressPackageTargetFrameworkCompatibility Include="$(UAPvNextTFM)" />
     <ProjectReference Include="..\..\System.Security.Principal.Windows\ref\System.Security.Principal.Windows.csproj" />
   </ItemGroup>
+  <ItemGroup Condition="'$(TargetGroup)' == 'netcoreapp'">
+    <ProjectReference Include="..\..\System.Runtime\ref\System.Runtime.csproj" />
+    <ProjectReference Include="..\..\System.Collections.NonGeneric\ref\System.Collections.NonGeneric.csproj" />
+  </ItemGroup>
 </Project>
