+++ src/System.Security.Permissions/ref/Configurations.props
@@ -8,6 +8,7 @@
     <BuildConfigurations>
       $(PackageConfigurations);
       netfx;
+      netcoreapp;
     </BuildConfigurations>
   </PropertyGroup>
 </Project>
