+++ src/System.Security.Permissions/src/Configurations.props
@@ -7,6 +7,7 @@
     </PackageConfigurations>
     <BuildConfigurations>
       $(PackageConfigurations);
+      netcoreapp;
       netfx;
     </BuildConfigurations>
   </PropertyGroup>
