+++ src/System.Security.Principal.Windows/ref/Configurations.props
@@ -4,6 +4,7 @@
     <BuildConfigurations>
       netstandard;
       netfx;
+      netcoreapp;
       net461;
     </BuildConfigurations>
   </PropertyGroup>
