+++ src/System.Security.Principal.Windows/ref/System.Security.Principal.Windows.csproj
@@ -15,4 +15,9 @@
     <Reference Include="mscorlib" />
     <Reference Include="System" />
   </ItemGroup>
+  <ItemGroup Condition="'$(TargetGroup)' == 'netcoreapp'">
+    <ProjectReference Include="..\..\System.Runtime\ref\System.Runtime.csproj" />
+    <ProjectReference Include="..\..\System.Security.Claims\ref\System.Security.Claims.csproj" />
+    <ProjectReference Include="..\..\System.Security.Principal\ref\System.Security.Principal.csproj" />
+  </ItemGroup>
 </Project>
